Implementacion del service

+++ src/main/java/com/cesar/dragonball/backend/api/repository/UniversoRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface UniversoRepository extends JpaRepository<Universo, Integer> {
+public interface UniversoRepository extends JpaRepository<Universo, Long> {
 }
